Revert "feat: add DISKANN indexes and the F16 vector element type (#163)"
This reverts commit 45ae099192a2a25aaa548ba67d7434e488a0c436.
